fix(android/engine): Refactor BaseActivity to KMEA
