+++ android/KMEA/app/build.gradle
@@ -66,6 +66,7 @@ dependencies {
     implementation 'com.google.android.material:material:1.2.1'
     implementation 'commons-io:commons-io:2.6'
     implementation 'io.sentry:sentry-android:3.1.0'
+    implementation 'androidx.preference:preference:1.1.1'
 
     // Robolectric
     testImplementation 'androidx.test:core:1.2.0'
