+++ android/KMEA/app/src/main/java/com/tavultesoft/kmea/BaseActivity.java
@@ -2,7 +2,7 @@
  * Copyright (C) 2021 SIL International. All rights reserved.
  */
 
-package com.tavultesoft.kmapro;
+package com.tavultesoft.kmea;
 
 import android.content.Context;
 import android.content.ContextWrapper;
@@ -10,6 +10,8 @@ import android.content.SharedPreferences;
 
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.preference.PreferenceManager;
+
+import com.tavultesoft.kmea.util.ContextUtils;
 
 import java.util.Locale;
 
@@ -20,7 +22,7 @@ public class BaseActivity extends AppCompatActivity {
     // Override the app locale using the BCP 47 tag from shared preferences
     // Using PreferenceManager because this is before onCreate()
     SharedPreferences prefs = PreferenceManager.getDefaultSharedPreferences(newBase);
-    String languageTag  = prefs.getString(KeymanSettingsActivity.displayLanguageKey, DisplayLanguages.unspecifiedLocale);
+    String languageTag  = prefs.getString(DisplayLanguages.displayLanguageKey, DisplayLanguages.unspecifiedLocale);
     if (languageTag == null || languageTag.equals(DisplayLanguages.unspecifiedLocale)) {
       // Use the default locale
       super.attachBaseContext(newBase);
