+++ android/KMEA/app/src/main/java/com/tavultesoft/kmea/ModelPickerActivity.java
@@ -40,7 +40,7 @@ import java.util.Map;
  * Gets the list of installable lexical models from Keyman cloud and allows user to download a model.
  * Displays a list of available models for a language ID.
  */
-public final class ModelPickerActivity extends AppCompatActivity {
+public final class ModelPickerActivity extends BaseActivity {
 
   private Context context;
   private static Toolbar toolbar = null;
